+++ src/main/java/roomescape/dto/InputValidator.java
@@ -2,9 +2,9 @@ package roomescape.dto;
 
 import roomescape.dto.exception.InputNotAllowedException;
 
-class InputValidator {
+public class InputValidator {
 
-    static void validateNotNull(Object... inputs) {
+    public static void validateNotNull(Object... inputs) {
         for (Object input : inputs) {
             if (input == null) {
                 throw new InputNotAllowedException("null은 입력할 수 없습니다.");
@@ -12,7 +12,7 @@ class InputValidator {
         }
     }
 
-    static void validateNotBlank(String... inputs) {
+    public static void validateNotBlank(String... inputs) {
         for (String input : inputs) {
             if (input.isBlank()) {
                 throw new InputNotAllowedException("빈 문자열은 입력할 수 없습니다.");
