+++ src/main/java/roomescape/dto/request/AdminReservationRequest.java
@@ -1,6 +1,8 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.time.LocalDate;
+
+import roomescape.dto.InputValidator;
 
 public record AdminReservationRequest(Long memberId, LocalDate date, Long timeId, Long themeId) {
