+++ src/main/java/roomescape/dto/request/MemberReservationRequest.java
@@ -1,6 +1,8 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import java.time.LocalDate;
+
+import roomescape.dto.InputValidator;
 
 public record MemberReservationRequest(LocalDate date, Long timeId, Long themeId) {
