+++ src/main/java/roomescape/dto/response/AvailableReservationTimeResponse.java
@@ -1,8 +1,8 @@
-package roomescape.dto;
-
-import roomescape.domain.ReservationTime;
+package roomescape.dto.response;
 
 import java.time.LocalTime;
+
+import roomescape.domain.ReservationTime;
 
 public record AvailableReservationTimeResponse(Long timeId, LocalTime startAt, boolean alreadyBooked) {
