+++ src/main/java/roomescape/dto/response/MemberPreviewResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.response;
 
 import roomescape.domain.Member;
