+++ src/main/java/roomescape/dto/response/ReservationTimeResponse.java
@@ -1,8 +1,8 @@
-package roomescape.dto;
-
-import roomescape.domain.ReservationTime;
+package roomescape.dto.response;
 
 import java.time.LocalTime;
+
+import roomescape.domain.ReservationTime;
 
 public record ReservationTimeResponse(Long id, LocalTime startAt) {
