+++ src/test/java/roomescape/acceptance/admin/ReservationAcceptanceTest.java
@@ -1,26 +1,35 @@
 package roomescape.acceptance.admin;
 
-import io.restassured.RestAssured;
-import io.restassured.common.mapper.TypeRef;
-import io.restassured.http.ContentType;
-import io.restassured.response.ValidatableResponse;
-import org.junit.jupiter.api.*;
-import org.springframework.http.HttpStatus;
-import roomescape.acceptance.BaseAcceptanceTest;
-import roomescape.acceptance.NestedAcceptanceTest;
-import roomescape.controller.exception.CustomExceptionResponse;
-import roomescape.dto.AdminReservationRequest;
-import roomescape.dto.ReservationResponse;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.Matchers.containsString;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+import static roomescape.acceptance.Fixture.adminToken;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_CUSTOMER_1;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_RESERVATION_1;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_RESERVATION_TIME_1;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_THEME_1;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Stream;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.containsString;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static roomescape.acceptance.Fixture.adminToken;
-import static roomescape.acceptance.PreInsertedData.*;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.DynamicTest;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestFactory;
+import org.springframework.http.HttpStatus;
+
+import io.restassured.RestAssured;
+import io.restassured.common.mapper.TypeRef;
+import io.restassured.http.ContentType;
+import io.restassured.response.ValidatableResponse;
+import roomescape.acceptance.BaseAcceptanceTest;
+import roomescape.acceptance.NestedAcceptanceTest;
+import roomescape.controller.exception.CustomExceptionResponse;
+import roomescape.dto.request.AdminReservationRequest;
+import roomescape.dto.response.ReservationResponse;
 
 class ReservationAcceptanceTest extends BaseAcceptanceTest {
