+++ src/test/java/roomescape/acceptance/admin/ReservationFilterAcceptanceTest.java
@@ -1,18 +1,23 @@
 package roomescape.acceptance.admin;
 
-import io.restassured.RestAssured;
-import io.restassured.common.mapper.TypeRef;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import roomescape.acceptance.BaseAcceptanceTest;
-import roomescape.dto.ReservationResponse;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import static roomescape.acceptance.Fixture.adminToken;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_RESERVATION_1;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_RESERVATION_2;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_RESERVATION_3;
+import static roomescape.acceptance.PreInsertedData.PRE_INSERTED_RESERVATION_4;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static roomescape.acceptance.Fixture.adminToken;
-import static roomescape.acceptance.PreInsertedData.*;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+
+import io.restassured.RestAssured;
+import io.restassured.common.mapper.TypeRef;
+import roomescape.acceptance.BaseAcceptanceTest;
+import roomescape.dto.response.ReservationResponse;
 
 @DisplayName("관리자가 예약을 필터링해서 조회한다.")
 class ReservationFilterAcceptanceTest extends BaseAcceptanceTest {
