+++ src/test/java/roomescape/acceptance/guest/ReservationTimeAcceptanceTest.java
@@ -16,7 +16,7 @@ import org.springframework.http.HttpStatus;
 import io.restassured.RestAssured;
 import io.restassured.common.mapper.TypeRef;
 import roomescape.acceptance.BaseAcceptanceTest;
-import roomescape.dto.AvailableReservationTimeResponse;
+import roomescape.dto.response.AvailableReservationTimeResponse;
 
 class ReservationTimeAcceptanceTest extends BaseAcceptanceTest {
